Restruct srvs -- moving master/*, aggregator/*., and worker/* to srvs/
