+++ srvs/gibbs.go
@@ -1,0 +1,4 @@
+package srvs
+
+func (wf *Master) Gibbs(iter int) {
+}
