+++ srvs/recover_test.go
@@ -1,4 +1,4 @@
-package master
+package srvs
 
 import (
 	"testing"
